Make FileOutStream an abstract class similar to FileInStream.
